backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-json-logger==2.0.7
structlog==23.2.0
aiofiles==23.2.1
pypdf==4.2.0
python-docx==1.1.2
httpx==0.27.2
sentence-transformers==3.2.1
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
fpdf2==2.7.9
python-dotenv==1.0.0
cohere==5.20.0
